refactor(graphql): isolate demo question seeding
